refactor(sparks): cleanup sparks code
